requirements.txt
flask==2.0.1
python-dotenv==0.19.0
transformers==4.31.0
torch==2.0.1
requests==2.31.0
